+++ src/components/security/BroadcastDialog.tsx
@@ -5,10 +5,10 @@ import { Card, CardContent } from "@/components/ui/card"
 import { Loader2, Radio, Network, ArrowUpRight, InfoIcon } from "lucide-react"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip"
 import { Badge } from "@/components/ui/badge"
-import { TxInfoCard } from "./TxInfoCard"
+import { TxInfoCard } from "../TxInfoCard"
 import { useState, useEffect } from "react"
 import { formatAddress } from "@/lib/utils"
-import { TxStatus } from "../particle-core/sdk/typescript/types/lib.index"
+import { TxStatus } from "../../particle-core/sdk/typescript/types/lib.index"
 import { Address } from "viem"
 import { CoreOperationType } from "@/types/OperationRegistry"
