+++ src/components/security/TemporalActionDialog.tsx
@@ -7,15 +7,15 @@ import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs"
 import { Card, CardContent } from "@/components/ui/card"
 import { Loader2, X, CheckCircle2, Clock, Shield, Wallet } from "lucide-react"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip"
-import { TxRecord } from "../particle-core/sdk/typescript/interfaces/lib.index"
+import { TxRecord } from "../../particle-core/sdk/typescript/interfaces/lib.index"
 import { formatAddress } from "@/lib/utils"
 import { Progress } from "@/components/ui/progress"
 import { useState, useEffect, FormEvent } from "react"
-import { TxInfoCard } from "./TxInfoCard"
+import { TxInfoCard } from "../TxInfoCard"
 import { cn } from "@/lib/utils"
 import { Badge } from "@/components/ui/badge"
 import { useWorkflowManager } from "@/hooks/useWorkflowManager"
-import { CoreOperationType, OperationPhase } from "../types/OperationRegistry"
+import { CoreOperationType, OperationPhase } from "../../types/OperationRegistry"
 import { Address } from "viem"
 import { toast } from "@/components/ui/use-toast"
